# backend/docs/group-chat-swagger.yaml
openapi: 3.0.3
info:
  title: 智慧共享单车调度系统 - 群聊功能API
  description: 完整的群聊功能API文档，包括群聊管理、消息传输、用户管理、WebSocket实时通信等核心功能
  version: 1.0.0
  contact:
    name: 开发团队
    email: support@bikehub.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api/chat
    description: 开发环境
  - url: https://api.bikehub.com/api/chat
    description: 生产环境
...
paths:
  /groups:
    get:
      tags:
        - 群聊管理
      summary: 获取用户群聊列表
      description: 获取当前用户所属的群聊列表，支持分页
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: 页码，默认为1
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          description: 每页数量，默认20，最大100
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: 成功获取群聊列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  groups:
                    type: array
                    items:
                      $ref: '#/components/schemas/GroupInfo'
                  total:
                    type: integer
                    description: 总群聊数量
                  page:
                    type: integer
                    description: 当前页码
                  page_size:
                    type: integer
                    description: 每页数量
                  has_more:
                    type: boolean
                    description: 是否有更多数据
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - 群聊管理
      summary: 创建群聊
      description: 创建一个新的群聊
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - group_type
              properties:
                name:
                  type: string
                  description: 群聊名称
                  example: 新群聊
                description:
                  type: string
                  description: 群聊描述
                  example: 群聊描述
                group_type:
                  $ref: '#/components/schemas/GroupType'
                max_members:
                  type: integer
                  minimum: 2
                  maximum: 500
                  default: 100
                  description: 最大成员数
      responses:
        '201':
          description: 群聊创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 群聊创建成功
                  group:
                    $ref: '#/components/schemas/Group'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /groups/{group_id}:
    get:
      tags:
        - 群聊管理
      summary: 获取群聊详情
      description: 获取指定群聊的详细信息
      security:
        - bearerAuth: []
      parameters:
        - name: group_id
          in: path
          required: true
          description: 群聊ID
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: 成功获取群聊详情
          content:
            application/json:
              schema:
                type: object
                properties:
                  group:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: 运维交流群
                      description:
                        type: string
                        example: 运维人员交流群
                      group_type:
                        $ref: '#/components/schemas/GroupType'
                      max_members:
                        type: integer
                        example: 100
                      member_count:
                        type: integer
                        example: 15
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - 群聊管理
      summary: 更新群聊信息
      description: 更新指定群聊的信息（仅群主和管理员）
      security:
        - bearerAuth: []
      parameters:
        - name: group_id
          in: path
          required: true
          description: 群聊ID
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 更新后的群名
                  example: 更新后的群名
                description:
                  type: string
                  description: 更新后的描述
                  example: 更新后的描述
                max_members:
                  type: integer
                  minimum: 2
                  maximum: 500
                  description: 最大成员数
                  example: 150
      responses:
        '200':
          description: 群聊信息更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 群聊信息更新成功
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - 群聊管理
      summary: 删除群聊
      description: 删除指定的群聊（仅群主）
      security:
        - bearerAuth: []
      parameters:
        - name: group_id
          in: path
          required: true
          description: 群聊ID
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: 群聊删除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 群聊删除成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /messages/{message_id}/read:
    post:
      tags:
        - 消息管理
      summary: 标记消息为已读
      description: 将指定消息标记为已读状态
      security:
        - bearerAuth: []
      parameters:
        - name: message_id
          in: path
          required: true
          description: 消息ID
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: 消息已标记为已读
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 消息已标记为已读
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /statistics:
    get:
      tags:
        - 统计信息
      summary: 获取聊天统计
      description: 获取用户的聊天统计信息
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 统计信息获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_groups:
                    type: integer
                    description: 总群聊数量
                    example: 5
                  total_messages:
                    type: integer
                    description: 总消息数量
                    example: 1500
                  unread_messages:
                    type: integer
                    description: 未读消息数量
                    example: 12
                  active_groups:
                    type: integer
                    description: 活跃群聊数量
                    example: 4
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /admin/groups:
    get:
      tags:
        - 管理员功能
      summary: 管理员获取所有群聊
      description: 管理员获取系统中所有群聊列表
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: 页码，默认为1
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          description: 每页数量，默认20，最大100
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: 成功获取所有群聊列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  groups:
                    type: array
                    items:
                      $ref: '#/components/schemas/Group'
                  total:
                    type: integer
                    description: 总群聊数量
                  page:
                    type: integer
                    description: 当前页码
                  page_size:
                    type: integer
                    description: 每页数量
                  has_more:
                    type: boolean
                    description: 是否有更多数据
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    GroupInfo:
      type: object
      properties:
        id:
          type: integer
          description: 群聊ID
          example: 1
        name:
          type: string
          description: 群聊名称
          example: 运维交流群
        description:
          type: string
          description: 群聊描述
          example: 运维人员交流群
        avatar_url:
          type: string
          nullable: true
          description: 群聊头像URL
        group_type:
          $ref: '#/components/schemas/GroupType'
        max_members:
          type: integer
          description: 最大成员数
          example: 100
        created_by:
          type: integer
          description: 创建者用户ID
          example: 1
        is_active:
          type: boolean
          description: 群聊是否激活
          example: true
        created_at:
          type: string
          format: date-time
          description: 创建时间
          example: "2025-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: 更新时间
          example: "2025-01-15T10:30:00Z"
        member_count:
          type: integer
          description: 当前成员数量
          example: 15
        last_message_time:
          type: string
          format: date-time
          nullable: true
          description: 最后消息时间
          example: "2025-01-15T15:20:00Z"
        user_role:
          $ref: '#/components/schemas/MemberRole'
          description: 当前用户在该群聊中的角色
        is_muted:
          type: boolean
          description: 当前用户是否被禁言
          example: false
        last_read_at:
          type: string
          format: date-time
          nullable: true
          description: 当前用户最后阅读时间
          example: "2025-01-15T15:00:00Z"
        unread_count:
          type: integer
          description: 未读消息数量
          example: 3
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer Token 认证
...